fix: widen order type to float to support infinity values

+++ strands-py/src/strands/agent/agent.py
@@ -730,7 +730,7 @@ class Agent(AgentBase):
         callback: HookCallback[TEvent],
         event_type: type[TEvent] | list[type[TEvent]] | None = None,
         *,
-        order: int = HookOrder.DEFAULT,
+        order: float = HookOrder.DEFAULT,
     ) -> None:
         """Register a callback function for a specific event type.
