+++ strands-py/src/strands/multiagent/base.py
@@ -256,7 +256,7 @@ class MultiAgentBase(ABC):
         raise NotImplementedError
 
     def add_hook(
-        self, callback: HookCallback, event_type: type | list[type] | None = None, *, order: int = HookOrder.DEFAULT
+        self, callback: HookCallback, event_type: type | list[type] | None = None, *, order: float = HookOrder.DEFAULT
     ) -> None:
         """Register a hook callback with the orchestrator.
